requirements.txt
asyncio 
aiohttp==1.0.5
discord.py==0.16.12 
